+++ src/constants.js
@@ -155,3 +155,6 @@ export const LONG_CLICK_TIME = 500;
 export const Z_NAMES = new Set(['z', 'valueZ', 'valueA', 'value1', 'category']);
 export const W_NAMES = new Set(['w', 'valueW', 'valueB', 'value2', 'value']);
 export const DEFAULT_IMAGE_LOAD_TIMEOUT = 15000;
+
+// Error messages
+export const ERROR_POINTS_NOT_DRAWN = 'Points have not been drawn';